Got matchbox_super_client to compile to WASM. Going to test using regular full mesh architecture
