--- matchbox_super_client/src/main.rs
@@ -1,5 +1,0 @@
-mod client;
-
-fn main() {
-    println!("Hello, world!");
-}
